moved code to proper namespace
