+++ api/src/main/java/cz/fi/muni/pa165/api/facade/RentApplicationFacadeImpl.java
@@ -1,4 +1,4 @@
-package cz.fi.muni.pa165.Api;
+package cz.fi.muni.pa165.api.facade;
 
 import cz.fi.muni.pa165.Api.DTO.RentApplicationDTO;
 import cz.fi.muni.pa165.Model.Entity.RentApplication;
